refactor(characters): move materials in separate folder
