+++ src/charactersData/characters/AratakiItto.ts
@@ -6,7 +6,7 @@ import {
   Gem,
   Materials,
   OtherMaterials
-} from '../materialNames';
+} from '../materials/materialNames';
 import {fillAscensionMaterials, fillTalentMaterials} from '../../utils/utils';
 
 const AratakiItto = {
