+++ src/charactersData/characters/Eula.ts
@@ -6,7 +6,7 @@ import {
   Flower,
   Books,
   BossSkillMaterial
-} from '../materialNames';
+} from '../materials/materialNames';
 import {fillAscensionMaterials, fillTalentMaterials} from '../../utils/utils';
 
 const Eula = {
