+++ src/charactersData/characters/Fischl.ts
@@ -6,7 +6,7 @@ import {
   Flower,
   Gem,
   Materials
-} from '../materialNames';
+} from '../materials/materialNames';
 import {fillAscensionMaterials, fillTalentMaterials} from '../../utils/utils';
 
 const Fischl = {
