+++ src/charactersData/materials/interactiveMapLinks.ts
