+++ src/charactersData/materials/materialNames.ts
